fixed a few compile warnings. fixed the eol-style on all the cpp and .h files.

+++ trunk/delta3d/utilities/AnimationViewer/Viewer.cpp
@@ -183,13 +183,13 @@ void Viewer::OnLoadCharFile( const QString &filename )
       QString nameToSend = QString::fromStdString(wrapper->GetCoreMaterialName(matID));
 
       osg::Vec4 diffuse = wrapper->GetCoreMaterialDiffuse(matID);
-      QColor diffColor(diffuse[0], diffuse[1], diffuse[2], diffuse[3]);
+      QColor diffColor( (int)(diffuse[0]), (int)(diffuse[1]), (int)(diffuse[2]), (int)(diffuse[3]) );
 
       osg::Vec4 ambient = wrapper->GetCoreMaterialAmbient(matID);
-      QColor ambColor(ambient[0], ambient[1], ambient[2], ambient[3]);
+      QColor ambColor( (int)(ambient[0]), (int)(ambient[1]), (int)(ambient[2]), (int)(ambient[3]) );
 
       osg::Vec4 specular = wrapper->GetCoreMaterialSpecular(matID);
-      QColor specColor(specular[0], specular[1], specular[2], specular[3]);
+      QColor specColor( (int)(specular[0]), (int)(specular[1]), (int)(specular[2]), (int)(specular[3]) );
 
       float shininess = wrapper->GetCoreMaterialShininess(matID);
